Add tests for list task presenter unit tests
